week-5/RAG_SupplyChain/requirements.txt
# Core dependencies
streamlit>=1.24.0
python-dotenv>=0.19.0
openai>=1.0.0
PyMuPDF>=1.22.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
stqdm>=0.0.5

# Optional
watchdog>=3.0.0  # Streamlit dev experience 